Move the tracking of values into instances
Rather than maintain the mapping of instances to values inside of the
`Preloader`, this commit moves the tracking into the individual
instances.

This is nice in the case that an object is used with multiple batches,
but also makes the code much clearer to reason about.

Note: If an object is run in two separate batches with the same method
name, it'll re-run the fetch call. This can be fixed in a follow-on PR.

Co-authored-by: John Hawthorn <john@hawthorn.email>