Add ability to define an inverse method
Currently, prelude requires that you define/call batch methods on the instances that would be the "N" part of an N+1 query.

Example:

```ruby
repositories.each.with_prelude do |repo|
  repo.starred_by?(current_user)
end
```

Sometimes it may feel more natural/elegant to flip this around:

```ruby
repositories.each.with_prelude do |repo|
  current_user.has_starred?(repo)
end
```

Another reason to have the inverse method could be to reduce churn in an existing large codebase where the "inverse" form of the method already has hundreds of callsites.

Why not just define the inverse method manually? That's a fair question. I think the only real benefit of this is the logical co-location of the inverse method with the batch method definition. On one hand, finding the definition of `User#has_starred?` may be harder. On the other hand, once you grep for `:has_starred?` you'll find it along with its inverse batch method definition, which is effectively the definition of the inverse method as well.

That all being said, there are some real tradeoffs here, and I'm not sure myself if the beneifts outweight the costs. Either way, this was fun to hack on and I learned more about prelude internals!