Add strict form of each prelude method that forbids N+1s
It's actually surprisingly possible in real world code to end up with an
array mixed from several preloaded sources and some sources that are not
preloaded at all, which can trigger N+1s on apparently preloaded data.

To make it possible to avoid this, add a strict form of each prelude
method suffixed with `!` which raises
`Prelude::StrictLoadingViolationError` if it is called in a way that
would otherwise lazily inject a Preloader instance.